# [WIP] Use `.update_attribute` over `.save!` for User login