@kbrock 2020-08-24
> Yeah, so this is about reducing queries, where in a situation where there is actually database latency (in other words: not my laptop where these benchmarks were taken), this could have some improvements.

Sure wish we didn't have to give this disclaimer so often.

It is tricky for people to see numbers and then try and guess what the numbers would be if something like network latency were applied.