@Fryguy 2020-08-24
It's funny but I am frequently the one giving that disclaimer, too...from the review end though, I usually read the statistics given to me, and don't extrapolate for some reason 😄 

For testing latency (or perhaps to enhance your number when queries are involved), it might be worth it to check out the `tc` tool, which allows you to introduce latency on your network connection.