@kbrock 2020-08-24
For a while I was running the postgres daemon in a container (the database was still on my main machine)
But just the transfer over the virtual network really showed the difference this latency can make.

Of course it got corrupted and was enough of a pain to get back up and running that I ended back in bare metal land