@NickLaMuro 2020-08-24
> Don't know why this popped in my head at 12:45 am, but it is what it is 😆

I saw this come in on Friday, and was thinking it was a little weird myself 😄 

Anyway, I like the idea, so I will more this `[WIP]` and test it out at some point and confirm the same number of queries with that change.