@NickLaMuro 2020-08-24
Early tests still show 24 queries being executed on the `/api/auth` route... and this was what I tried to change:

```diff
diff --git a/lib/unique_within_region_validator.rb b/lib/unique_within_region_validator.rb
index ee90e944f5..4bae38c61b 100644
--- a/lib/unique_within_region_validator.rb
+++ b/lib/unique_within_region_validator.rb
@@ -14,6 +14,8 @@
 class UniqueWithinRegionValidator < ActiveModel::EachValidator
   def validate_each(record, attribute, value)
     return if value.nil?
+    return unless record.changed.include?(attribute)
+
     match_case = options.key?(:match_case) ? options[:match_case] : true
     record_base_class = record.class.base_class
     matches =
```

(did also "revert" the `.update_attribute` calls as well)

Still haven't figured out what are causing the extra calls yet.