@NickLaMuro 2020-08-24
Okay, so I found our culprit:

https://github.com/ManageIQ/manageiq/blob/f0b6fa0a07484d80748f9c47df059596502b7bd3/app/models/user.rb#L47

This validator causes the following two queries after each save:

```SQL
SELECT  "miq_groups".* FROM "miq_groups" WHERE "miq_groups"."id" = $1 LIMIT $2

SELECT  1 AS one
FROM "miq_groups"
INNER JOIN "miq_groups_users" ON "miq_groups"."id" = "miq_groups_users"."miq_group_id"
WHERE "miq_groups_users"."user_id" = $1
  AND "miq_groups"."id" = $2
LIMIT $3
```

So making the above change to our custom validator removes two out of the extra 6 queries, but doesn't do the ones from this built in validator.