@Fryguy 2020-08-25
Wonder if there's a way to make that `validates_inclusion_of  :current_group` nicer (perhaps also similarly for changes to `current_group`)