@NickLaMuro 2020-09-02
> Wonder if there's a way to make that `validates_inclusion_of :current_group` nicer (perhaps also similarly by checking there are changes to `current_group`)

@Fryguy Is this something that we want to try for now and in lieu of these changes, or is this just a thought for the future?  If it is the latter, then I am going to remove the `[WIP]` from this PR then.