@kbrock 2020-09-03
```diff
-    return unless record.changed.include?(attribute)
+    return unless record.changed.include?(attribute.to_s)
```