@Fryguy 2020-09-09
> @Fryguy Is this something that we want to try for now and in lieu of these changes, or is this just a thought for the future? If it is the latter, then I am going to remove the `[WIP]` from this PR then.

I'm thinking now.  It seems to make sense, as checking changed would help validations across the board, and my hunch is that if we get rid of the validations, then the update_attribute change would no longer be necessary.