@NickLaMuro 2020-09-09
Cool.  Going to leave this open as until that effort is complete as a reminder (mostly to myself) to check once those changes are merge that the new validator had the same effect.