@kbrock 2020-09-12
Hope these will reduce the number of validations run:

- https://github.com/ManageIQ/manageiq/pull/20520 (followup: https://github.com/ManageIQ/manageiq/pull/20539)
- https://github.com/ManageIQ/manageiq/pull/20526